# Conflicts:
#	src/main/java/com/VietBlog/controller/BaiVietController.java
#	src/main/java/com/VietBlog/entity/BaiViet.java
#	src/main/java/com/VietBlog/entity/User.java
#	src/main/java/com/VietBlog/repository/BaiVietRepository.java
#	src/main/java/com/VietBlog/service/UserService.java
#	src/main/resources/application.properties

+++ src/main/java/com/VietBlog/repository/UserRepository.java
@@ -1,14 +1,13 @@
 package com.VietBlog.repository;
 
-import java.util.Optional;
-
+import com.VietBlog.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.VietBlog.entity.User;
+import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
     Optional<User> findByDienThoai(String dienThoai);
 }
